bdpm/requirements.txt
django
django-rest-framework
django-extensions
psycopg2
